requirements.txt
---
### requirements.txt
```

plaintext
python-dotenv
ccxt
telethon
beautifulsoup4
selenium
tweepy
requests
